fix resource compare bug

+++ src/sdk/galaxy.h
@@ -59,7 +59,7 @@ struct JobInformation {
     int32_t priority;
     int32_t running_num;
     int32_t cpu_used;
-    int32_t mem_used;
+    int64_t mem_used;
     int32_t pending_num;
     int32_t deploying_num;
     std::string state;
